pyproject.toml
[tool.poetry]
name = "photo-share"
version = "0.1.0"
description = ""
authors = ["ivcoder <vank300828@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
psycopg2-binary = "^2.9.10"
alembic = "^1.15.2"
pydantic-settings = "^2.9.1"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.11.4"}
passlib = "^1.7.4"
python-jose = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
